feat: auth.js sign.js 모듈화
