+++ pages/signin.html
@@ -8,7 +8,7 @@
 		<link rel="stylesheet" href="/styles/variables.css" />
 		<link rel="stylesheet" href="/styles/auth.css" />
 		<title>Linkbrary</title>
-		<script src="/utils/auth.js" defer></script>
+		<script type="module" src="/utils/signin.js" defer></script>
 	</head>
 	<body>
 		<!-- H1 (a11y) -->
